refactor: move sample code to ff4j-spring-boot-samples with modules
