--- ff4j-spring-boot-sample/src/main/java/org/ff4j/sample/config/CORSConfig.java
@@ -1,14 +1,0 @@
-package org.ff4j.sample.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class CORSConfig implements WebMvcConfigurer {
-
-  @Override
-  public void addCorsMappings(CorsRegistry registry) {
-    registry.addMapping("/**");
-  }
-}
