+++ ff4j-spring-boot-samples/ff4j-spring-boot-starter-webmvc-sample/src/main/java/org/ff4j/spring/boot/starter/webmvc/sample/config/CORSConfig.kt
@@ -1,0 +1,12 @@
+package org.ff4j.spring.boot.starter.webmvc.sample.config
+
+import org.springframework.context.annotation.Configuration
+import org.springframework.web.servlet.config.annotation.CorsRegistry
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
+
+@Configuration
+class CORSConfig : WebMvcConfigurer {
+    override fun addCorsMappings(registry: CorsRegistry) {
+        registry.addMapping("/**")
+    }
+}
